requirements.txt
pandas
finnhub-python
requests
jupyter
json
flask